+++ src/util/Initializers.java
@@ -4,7 +4,7 @@ import java.util.Random;
 
 public class Initializers {
     public interface Supplier {
-        double get();
+        float get();
     }
 
     public static Supplier ZEROS() {
@@ -15,21 +15,21 @@ public class Initializers {
         return () -> 0;
     }
 
-    public static Supplier CONST(double constant) {
+    public static Supplier CONST(float constant) {
         return () -> constant;
     }
 
-    public static Supplier RAND(double amplitude) {
-        return () -> (Math.random() * 2 - 1) * amplitude;
+    public static Supplier RAND(float amplitude) {
+        return () -> (float) ((Math.random() * 2 - 1) * amplitude);
     }
 
-    public static Supplier RAND(double amplitude, double mean) {
-        return () -> (Math.random() * 2 - 1) * amplitude + mean;
+    public static Supplier RAND(float amplitude, float mean) {
+        return () -> (float) ((Math.random() * 2 - 1) * amplitude + mean);
     }
 
-    public static Supplier GAUSSIAN(double mean, double standardDeviation) {
+    public static Supplier GAUSSIAN(float mean, float standardDeviation) {
         Random fRandom = new Random();
-        return () -> fRandom.nextGaussian() * standardDeviation + mean;
+        return () -> (float) (fRandom.nextGaussian() * standardDeviation + mean);
     }
 
     /**
@@ -37,6 +37,6 @@ public class Initializers {
      * @param n Number of units in the current layer.
      */
     public static Supplier GLOROT(int n) {
-        return GAUSSIAN(0, 1 / Math.sqrt(n));
+        return GAUSSIAN(0f, (float) (1 / Math.sqrt(n)));
     }
 }
