+++ src/util/LossFunctions.java
@@ -2,11 +2,11 @@ package util;
 
 public class LossFunctions {
     public interface LossFn {
-        double f(double[] prediction, double[] truth);
+        float f(float[] prediction, float[] truth);
     }
 
     public static LossFn CATEGORICAL_CROSSENTROPY = (x, y) -> {
-        double crossEntropy = 0;
+        float crossEntropy = 0;
 
         for (int i = 0; i < y.length; i++) {
             crossEntropy -= y[i] * Math.log(x[i]);
@@ -16,7 +16,7 @@ public class LossFunctions {
     };
 
     public static LossFn MEAN_SQUARED_ERROR = (x, y) -> {
-        double sum = 0;
+        float sum = 0;
 
         for (int i = 0; i < y.length; i++) {
             sum += Math.pow(x[i] - y[i], 2);
